+++ Bundlr/src/Bundlr/Core/Constants.cs
@@ -6,7 +6,7 @@ namespace Bundlr.Core
         public const ushort HeaderSizeMax = 63;
         public const uint PayloadSizeMax = 256 * 1000;
         public const ushort HeaderItemByteLengthMax = 1023;
-        public const ushort HeaderItemDescriptorByteLength = sizeof(int);
+        public const ushort HeaderItemDescriptorByteLength = sizeof(ushort);
         public const ushort HeaderSizeDescriptorByteLength = sizeof(byte);
         public const ushort PayloadSizeDescriptorByteLength = sizeof(uint);
         public const ushort ChecksumSizeDescriptorByteLength = sizeof(ushort);
